その他のディレクトリもre-ducksパターンっぽい感じに
