+++ src/components/about/styles.tsx
