+++ src/components/top/styles.tsx
